Renamed `Rats`to `rats` as to conform to naming conventions. cleaned up imports
